*** empty log message ***

+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/service/BedManager.java
@@ -375,7 +375,7 @@ public class BedManager {
     			
     			if(beds[i].getRoomId().intValue() == rooms[j].getId().intValue()){
     				bedList.add(beds[i]);
-    				if(beds[i].isActive())rooms[j].setTotalBedOccupancy(rooms[j].getTotalBedOccupancy()+1);
+    				if(beds[i].isActive())rooms[j].setTotalBedOccupancy(Integer.valueOf(rooms[j].getTotalBedOccupancy().intValue()+1));
     				break;
     			}
     		}
